modify admin product insert function for product_manage

+++ portal/Controllers/php/file_2.php
@@ -1,10 +1,10 @@
 <?php
 session_start();
-	include('../model/php_model.php');
-	$or = new Orders();
-$action = $_GET['act'];
-if($_SESSION['ord_code'] !=""){ 
-	if($action=='delimg'){ //删除图片 
+include('../../model/php_model.php');
+$or = new Orders();
+$action = isset($_GET['act'])? $_GET['act'] : '';
+if(isset($_SESSION['ord_code']) && $_SESSION['ord_code'] !=""){
+    if($action=='delimg'){ //删除图片
 		$filename = $_POST['imagename']; 
 		if(!empty($filename)){ 
 			unlink('../admin/file/'.$_SESSION['shopping_user'][0]['memNo']."/".$filename);
